+++ tests/guardrails/test_traceloop_client.rs
@@ -1,5 +1,6 @@
 use hub_lib::guardrails::providers::traceloop::TraceloopClient;
-use hub_lib::guardrails::providers::{GuardrailClient, create_guardrail_client};
+use hub_lib::guardrails::providers::create_guardrail_client;
+use hub_lib::guardrails::types::GuardrailClient;
 use hub_lib::guardrails::types::GuardMode;
 use serde_json::json;
 use wiremock::matchers;
